# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    RefinementDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        entityId:
          type: string
        epicId:
          type: string
        projectId:
          type: string
        repositoryId:
          type: string
        branch:
          type: string
        parent:
          type: string
        label:
          type: string
        runtimeStatus:
          type: string
        runtimeError:
          type: string
        clean:
          type: boolean
        ahead:
          type: integer
          format: int32
        behind:
          type: integer
          format: int32
        conflictsWithParent:
          type: boolean
        agentActivity:
          type: string
        daemonConnectedAt:
          $ref: "#/components/schemas/Instant"
        daemonVersion:
          type: string
        daemonOutdated:
          type: boolean
        pinnedDaemonVersion:
          type: string
        daemonVersionStale:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Instant"
...
    WorkRefinementAnswer:
      type: object
      description: A work entity's refinement room; null on a read that found none.
      properties:
        refinement:
          $ref: "#/components/schemas/RefinementDto"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /projects/api/work/{qualifiedId}/refinement:
    post:
      summary: Open (or find) a work entity's refinement room
      description: "Find-or-create: an entity has at most one room, and a second open\
        \ answers the first. Refused for a feature or a task, an entity not at REPORTED,\
        \ one an ACTIVE workspace names, and a project with no wrapper. The path names\
        \ the entity by qualified id (<projectSlug>-<n>) or UUID."
      operationId: startWorkRefinement
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkRefinementAnswer"
        "404":
          description: No entity with this id
        "409":
          description: The entity cannot be refined now
        "502":
          description: The git host would not cut the branch
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Refinement Controller
      security:
      - SecurityScheme:
        - qits:admin
    get:
      summary: Find a work entity's refinement room
      description: "The entity's room, or null when it has none; never creates one.\
        \ The path names the entity by qualified id or UUID."
      operationId: getWorkRefinement
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "The room, or null"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkRefinementAnswer"
        "404":
          description: No entity with this id
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Refinement Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
...
info:
  title: qits-projects API
  version: 2026.1006.121753